requirements.txt
# Production requirements for Railway/Cloud deployment
# Use requirements_cloud.txt content

Flask==3.0.0
Werkzeug==3.0.1
opencv-python-headless==4.8.1.78
numpy==1.24.3
face-recognition==1.3.0
pyzbar==0.1.9
pywhatkit==5.4
gunicorn==21.2.0
Pillow==10.1.0
pymongo==4.6.0
dnspython==2.4.2
python-dotenv==1.0.0
